Moved NoExitSecurityManager to dpc-testing
