+++ dpc-testing/src/main/java/gov/cms/dpc/testing/NoExitSecurityManager.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.api;
+package gov.cms.dpc.testing;
 
 import java.security.Permission;
 
@@ -6,10 +6,10 @@ import java.security.Permission;
  * Use this to test classes that call System.exit() to prevent them from exiting early.
  * Before your test, call...
  * System.setSecurityManager(new NoExitSecurityManager())
- * and after make sure to reset it.
- * Instead of exiting when the class being tested calls System.exit(), an exception will be thrown.
+ * and after make sure to reset it.  Instead of exiting when the class being tested calls System.exit(), a
+ * RuntimeException will be thrown with its message set to the exit status.
  */
-class NoExitSecurityManager extends SecurityManager {
+public class NoExitSecurityManager extends SecurityManager {
     @Override
     public void checkPermission(Permission perm) {
         // Do nothing
@@ -17,6 +17,7 @@ class NoExitSecurityManager extends SecurityManager {
 
     @Override
     public void checkExit(int status) {
-        // Do nothing
+        super.checkExit(status);
+        throw new RuntimeException(String.valueOf(status));
     }
 }
